fix: enable modal switcher for other layouts other then Onboarding
