+++ packages/suite/src/views/start/SuiteStart.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 
 import { StartContent } from './StartContent';
-import { WelcomeLayout } from '../../components/suite/layouts/WelcomeLayout/WelcomeLayout';
+import { WelcomeLayoutWithoutModalSwitcher } from '../../components/suite/layouts/WelcomeLayout/WelcomeLayoutWithoutModalSwitcher';
 
 const Content = styled.div`
     display: flex;
@@ -11,9 +11,16 @@ const Content = styled.div`
 `;
 
 export const SuiteStart = () => (
-    <WelcomeLayout>
+    /**
+     * In onboarding we have custom confirm dialogs that rely on the fact,
+     * that we do not have the ModalProvider in layout, and therefore it is
+     * handled in a custom way in the onboarding.
+     *
+     * Go to `OnboardingStepBox` search for `ConfirmOnDevice`.
+     */
+    <WelcomeLayoutWithoutModalSwitcher>
         <Content data-testid="@onboarding/welcome">
             <StartContent />
         </Content>
-    </WelcomeLayout>
+    </WelcomeLayoutWithoutModalSwitcher>
 );
